requirements.txt
# Flask backend
Flask==3.1.2
flask-cors==6.0.1
gunicorn==23.0.0

# Core ML dependencies
numpy==1.26.4
scipy==1.12.0
scikit-learn==1.4.2
opencv-contrib-python==4.10.0.84
Pillow==10.3.0
joblib==1.4.2

# TensorFlow (CPU optimized for Render)
tensorflow-cpu==2.20.0
keras==3.10.0
h5py==3.11.0
protobuf==5.28.3

# Auth and Google integrations (if needed)
google-auth==2.33.0
google-auth-oauthlib==1.2.1

# Utils & formatting
Markdown==3.6
rich==13.8.0
requests==2.32.3